backend/requirements.txt
# Hea Early Health Risk Detector - Backend Dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
sqlalchemy>=2.0.0
alembic>=1.13.0
boto3>=1.34.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bleach>=6.0.0
python-multipart>=0.0.9
httpx>=0.26.0

# ML inference dependencies
transformers>=4.37.0
torch>=2.0.0
numpy>=1.26.0
scipy>=1.12.0
scikit-learn>=1.4.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Utilities
emoji>=2.10.0
textblob>=0.18.0
python-dateutil>=2.8.0

# Gemini AI
google-genai>=1.0.0
